+++ src/handler/lib/stream_merkleized_map_value.c
@@ -1,11 +1,13 @@
 #include "stream_merkleized_map_value.h"
+
+/* Local headers */
 #include "get_merkle_leaf_index.h"
 #include "stream_merkle_leaf_element.h"
 
 int call_stream_merkleized_map_value(dispatcher_context_t *dispatcher_context,
                                      const merkleized_map_commitment_t *map,
                                      const uint8_t *key,
-                                     int key_len,
+                                     size_t key_len,
                                      void (*len_callback)(size_t, void *),
                                      void (*callback)(buffer_t *, void *),
                                      void *callback_state) {
