+++ src/handler/lib/stream_merkleized_map_value.h
@@ -1,7 +1,8 @@
 #pragma once
 
-#include "../../boilerplate/dispatcher.h"
-#include "../../common/merkle.h"
+/* Local headers */
+#include "dispatcher.h"
+#include "merkle.h"
 
 /**
  * Given a commitment to a merkleized key-value map, this flow find out the index of the
@@ -16,7 +17,7 @@
 int call_stream_merkleized_map_value(dispatcher_context_t *dispatcher_context,
                                      const merkleized_map_commitment_t *map,
                                      const uint8_t *key,
-                                     int key_len,
+                                     size_t key_len,
                                      void (*len_callback)(size_t, void *),
                                      void (*callback)(buffer_t *, void *),
                                      void *callback_state);
